resolucion de conflictos en pantalla de Sign Up
