--- app/src/main/java/com/unlam/edu/ar/videotecamoviltp/Classes/Exceptions.kt
@@ -1,5 +1,0 @@
-package com.unlam.edu.ar.videotecamoviltp.Classes
-
-open class UserNotFoundException(msg : String) : Throwable() {
-
-}
