+++ app/src/main/java/com/unlam/edu/ar/videotecamoviltp/SignUpActivity.kt
@@ -3,6 +3,7 @@ package com.unlam.edu.ar.videotecamoviltp
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.util.Log
 import android.widget.Button
 import android.widget.EditText
 import android.widget.Toast
@@ -13,7 +14,7 @@ class SignUpActivity : AppCompatActivity() {
 
     companion object {
 
-        const val REGEX_EMAIL : String = "^[\\\\w!#\$%&’*+/=?`{|}~^-]+(?:\\\\.[\\\\w!#\$%&’*+/=?`{|}~^-]+)*@(?:[a-zA-Z0-9-]+\\\\.)+[a-zA-Z]{2,6}\$"
+        const val REGEX_EMAIL : String = "^[0-9a-zA-Z._.-]+\\@[0-9a-zA-Z._.-]+\\.[0-9a-zA-Z]+\$"
 
     }
 
@@ -21,7 +22,7 @@ class SignUpActivity : AppCompatActivity() {
     lateinit var email : EditText
     lateinit var password : EditText
     lateinit var signUp : Button
-    lateinit var listUsers : ListUsers
+    var listUsers : ListUsers = ListUsers()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -44,54 +45,53 @@ class SignUpActivity : AppCompatActivity() {
    private fun setListener() {
         signUp.setOnClickListener {
 
-            navigateToHome()
-          //  createUser()
+           createUser()
+
         }
     }
 
     private fun createUser() {
-        if(email.toString().isNotEmpty()
-            && name.toString().isNotEmpty() &&
-            password.toString().isNotEmpty()){
 
-            if(validateEmail(email)){
+        if(email.getText().toString().isNotEmpty()
+                && name.getText().toString().isNotEmpty()
+                        && password.getText().toString().isNotEmpty()){
 
-             if(listUsers.addUserToTheList(User(name = name.toString(),
-                                                email = email.toString(),
-                                                password = password.toString() ))){
-                 Toast.makeText(
-                     this@SignUpActivity,
-                     "Registración exitosa",
-                     Toast.LENGTH_LONG
-                 ).show()
-                 navigateToHome()
-             }else{
-                 Toast.makeText(
-                     this@SignUpActivity,
-                     "Usuario ya existente con email ${email}",
-                     Toast.LENGTH_LONG
-                 ).show()
-             }
-            }else{
-                Toast.makeText(
-                    this@SignUpActivity,
-                    "Formato de email incorrecto, pruebe con uno estilo juanperez@gmail.com",
-                    Toast.LENGTH_LONG
-                ).show()
-            }
-        }else{
-            Toast.makeText(
-                this@SignUpActivity,
-                "Por favor rellene los campos",
-                Toast.LENGTH_LONG
-            ).show()
-        }
+                                 if(validateEmail(email)){
+
+                                     if(listUsers.addUserToTheList(User(name = name.getText().toString().trim(),
+                                                                        email = email.getText().toString().trim(),
+                                                                        password = password.getText().toString() ))){
+                                                         Toast.makeText(
+                                                             this@SignUpActivity,
+                                                             "Registración exitosa",
+                                                             Toast.LENGTH_LONG
+                                                         ).show()
+                                                        navigateToHome()
+                                                     }else{
+                                                         Toast.makeText(
+                                                             this@SignUpActivity,
+                                                             "Usuario ya existente con email ${email.getText()}",
+                                                             Toast.LENGTH_LONG
+                                                         ).show()
+                                                     }
+                                                    }else{
+                                                        Toast.makeText(
+                                                            this@SignUpActivity,
+                                                            "Formato de email incorrecto, pruebe con uno estilo juanperez@gmail.com",
+                                                            Toast.LENGTH_LONG
+                                                        ).show()
+                                                    }
+                                                }else{
+                                                    Toast.makeText(
+                                                        this@SignUpActivity,
+                                                        "Por favor rellene todos los campos",
+                                                        Toast.LENGTH_LONG
+                                                    ).show()
+                                                }
     }
 
     private fun validateEmail(email: EditText): Boolean {
-            val regex = Regex(pattern = REGEX_EMAIL)
-            val matched = regex.containsMatchIn(email.toString().trim())
-            return matched
+        return email.getText().toString().trim().matches(Regex(pattern = REGEX_EMAIL))
     }
 
     private fun navigateToHome() {
